Update .env and security.yaml
In .env, the DATABASE\_URL for PostgreSQL is commented out.

In security.yaml, the user provider in the firewall 'main' is changed from 'users\_in\_memory' to 'app\_user\_provider', and 'app\_user\_provider' is configured to use the User entity with the property 'email'. The 'dev' firewall has security set to false.